# openapi.yaml
openapi: '3.1.0'
info:
    title: Manna AI Agent API
    version: '1.0.0'
    contact:
        name: Manna API
        url: https://github.com/Guebbit/manna
    description: |
        Local-first agentic AI platform. Exposes the agent reasoning loop, IDE-specific
        direct-LLM endpoints, file-upload helpers,
        and a multi-library PDF ingestion and semantic search system.

        **Base URL (default):** `http://localhost:3001`
servers:
    - url: http://localhost:3001
      description: Local development server
...
paths:
...
    /workflow:
        post:
            tags: [workflow]
            summary: Run an explicit ordered list of steps sequentially
            description: |
                Accepts an array of step task strings and runs each as an independent bounded
                `agent.run()` sub-call.  Each step has its own `maxStepsPerStep` iteration cap
                (defaulting to `AGENTS_MAX_STEPS`), so a misbehaving step cannot consume the
                entire budget.

                ### Context carry modes (`carry`)

                | Mode      | Behaviour |
                |-----------|-----------|
                | `none`    | Steps are fully isolated — no prior context forwarded. |
                | `summary` | A compact bullet-point summary of prior step results is prepended to each subsequent step prompt. **(default)** |
                | `full`    | The complete verbatim output of every prior step is appended. Context may grow quickly. |

                Use `POST /workflow` when your outer loop is already structured as an ordered list
                and you want each step bounded independently.
            operationId: postWorkflow
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/WorkflowRequest'
                        examples:
                            basic:
                                summary: Two-step workflow (default carry=summary)
                                value:
                                    steps:
                                        - task: 'List all TypeScript files under packages/'
                                        - task: 'Summarise the purpose of each listed file'
                            withOptions:
                                summary: Multi-step write workflow with per-step cap
                                value:
                                    steps:
                                        - task: 'Read the existing index.ts and summarise its structure'
                                        - task: 'Create a new helper module at src/utils/helper.ts'
                                        - task: 'Write unit tests for the helper module'
                                    allowWrite: true
                                    profile: code
                                    carry: full
                                    maxStepsPerStep: 10
            responses:
                '200':
                    description: 'All steps completed (some may have `success: false` if individual steps errored)'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/WorkflowResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalError'
    /workflow/stream:
        post:
            tags: [workflow]
            summary: Run a sequential workflow with SSE streaming
            description: |
                Same as `POST /workflow`, but streams lifecycle events as Server-Sent Events.
                The connection stays open until all steps complete.

                SSE event types (in order):
                - `workflow_start` — before the first step; `{ stepCount }`.
                - `step_start`     — before each step; `{ index, task }`.
                - `step`           — inner agent iteration; `{ workflowIndex, step, action, thought }`.
                - `tool`           — tool result/error; `{ workflowIndex, tool, result? | error? }`.
                - `route`          — model routing decision; `{ workflowIndex, profile, model, reason }`.
                - `step_done`      — after each step; full `WorkflowStepResult` object.
                - `done`           — after all steps; full `WorkflowResponse` object.
                - `error`          — on fatal error; `{ error }`.
            operationId: postWorkflowStream
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/WorkflowRequest'
            responses:
                '200':
                    description: SSE event stream
                    content:
                        text/event-stream:
                            schema:
                                type: string
                                description: |
                                    Server-Sent Events stream. Each event is typed
                                    (workflow_start, step_start, step, tool, route, step_done, done, error).
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalError'

    # ── IDE endpoints ─────────────────────────────────────────────────────────
...
components:
...
    responses:
        BadRequest:
            description: Invalid request body or parameters
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
...
        InternalError:
            description: Internal server error (LLM failure, Qdrant unavailable, etc.)
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
    schemas:
        ErrorResponse:
            type: object
            required: [error]
            properties:
                error:
                    type: string
                    description: Human-readable error message.
                    example: "Library 'scientific-american' not found"
...
        WorkflowRequest:
            type: object
            required: [steps]
            properties:
                steps:
                    type: array
                    items:
                        $ref: '#/components/schemas/StepDefinition'
                    minItems: 1
                    maxItems: 50
                    description: |
                        Ordered list of workflow step definitions.
                    example:
                        - task: 'List all TypeScript files under packages/'
                        - task: 'Summarise the purpose of each listed file'
                allowWrite:
                    type: boolean
                    default: false
                    description: |
                        When `true`, unlocks `write_file`, `scaffold_project`, and `document_ingest`
                        tools for all steps.
                profile:
                    type: string
                    enum: [fast, reasoning, code, default]
                    description: |
                        Force a specific model profile for all steps, bypassing automatic routing.
                carry:
                    type: string
                    enum: [none, summary, full]
                    default: summary
                    description: |
                        How prior step outputs are carried into subsequent steps.
                        `none` = isolated; `summary` = compact bullet summary (default);
                        `full` = verbatim full output.
                maxStepsPerStep:
                    type: integer
                    minimum: 1
                    maximum: 100
                    description: |
                        Per-step agent-loop iteration cap.  Overrides `AGENTS_MAX_STEPS` for this
                        request only.  Defaults to `AGENTS_MAX_STEPS` (env var, default 5).
        WorkflowStepResult:
            type: object
            properties:
                index:
                    type: integer
                    description: Zero-based position of this step in the workflow.
                task:
                    type: string
                    description: The original task string for this step.
                result:
                    type: string
                    description: The agent's final answer for this step.
                success:
                    type: boolean
                    description: Whether this step completed without throwing.
                durationMs:
                    type: number
                    description: Wall-clock execution time for this step in milliseconds.
                error:
                    type: string
                    description: Error message (only present when `success` is `false`).
        WorkflowResponse:
            type: object
            properties:
                steps:
                    type: array
                    items:
                        $ref: '#/components/schemas/WorkflowStepResult'
                    description: Per-step results in execution order.
                allSucceeded:
                    type: boolean
                    description: '`true` when every step completed successfully.'
                totalDurationMs:
                    type: number
                    description: Total wall-clock duration across all steps in milliseconds.

        # ── IDE schemas ───────────────────────────────────────────────────────────
...
        StepDefinition:
            type: object
            required: [task]
            properties:
                task:
                    type: string
                    minLength: 1
                    description: Task text to run for this workflow step.
            description: Workflow step input definition.
...